requirements.txt
streamlit
google-genai
edge-tts
Pillow
langdetect